Update to 1.7.6.2
This update fixes an incompatibility caused by the git push --quiet fix¹
included in 1.7.6.1.  Unfortunately, the "fix" was reverted.  A proper
fix for git push --quiet is being worked on for a future release.

Also worthy of note is that the source tarball for 1.7.6.2 is not
available due to the recent intrusion at kernel.org.  Instead, the
tarball was generated using 'make dist' from the git tree.  The v1.7.6.2
tag was verified using Junio's GPG key and compared against the 1.7.6.1
tarball to verify the diff.

¹ http://thread.gmane.org/gmane.comp.version-control.git/180652