Install git-prompt.sh in %{_sysconfdir}/profile.d
With git-1.7.12, __git_ps1 has moved to a separate script.  This is to
accommodate newer bash-completion which allows for dynamic loading of
completion files.

Having git-prompt.sh in profile.d ensures that the functions needed for
git in the prompt are present regardless of whether bash-completion is
installed or loaded.