Update asciidoc requirements, drop unsupported ASCIIDOC7
Since git-1.7.12, asciidoc >= 8.4.1 is a requirement to build the
documentation.  Update the build requires to reflect this.  Note that
this means that EL-5 can no longer build documentation as asciidoc-8.1.0
is the latest in EPEL.  (It might be worth adding support for using the
prebuilt documentation files to enable EL-5 builds to continue working.)