Update summary/description of numerous subpackages
Use '%{summary}.' as the %description for all subpackages where the
summary and description matched.  This is one less place to change
if/when we update a summary/description in the future.

While looking at rpmlint complaints, gitk had a spelling warning because
the summary/description used the British  English spelling of
"visualiser." Rather than simply change that to the American English
spelling, change the summary/description to match the gitk
documentation.  Do the same for git-gui.

Using %summary fixes the %description for perl-Git-SVN, which previously
used the %description from perl-Git mistakenly.

Also improve/clarify the summary of git-gnome-keyring, git-send-email,
and git-svn.