Improve test suite coverage
3041eeb (Run git test suite, 2016-12-16) enabled the test suite, but
many of the tests were skipped due to missing requirements.  Improve the
test suite coverage by ensuring as many requirements are present as
possible.

Also add an option to run the test suite with multiple jobs using the
%{?_smp_mflags} macro.  This cut the test suite run time roughly in
half in local testing.  The -O option to 'make' helps ensure the output
of each job is collected together rather than interspersed with output
from other jobs.  Sadly, the tests fail when run in parallel in koji, so
this is disabled by default.  It is included to speed up local builds.

To enable in rpmbuild and mock, use either "--with parallel_tests" or
"--define 'parallel_tests 1'" arguments.  This is only supported on
Fedora, as it requires make >= 4.0, which neither EL6 nor EL7 satisfy.

Disable 2 failing tests which run with a limited stack on aarch64, arm*,
and ppc* architectures.  Some debugging with the affected hardware is
needed.

Similarly, disable "git svn branch tests" globally for both
t9128-git-svn-cmd-branch and t9167-git-svn-cmd-branch-subproject because
they fail randomly.  The bug seems likely to be in the subversion perl
bindings, but requires more testing.  The failure output from a manual
run (./t9128-git-svn-cmd-branch.sh -v -x -d):

    ++ git svn branch a
    Copying file:///home/tmz/rpmbuild/BUILD/git-2.14.2/t/trash%20directory.t9128-git-svn-cmd-branch/svnrepo/trunk at r2 to file:///home/tmz/rpmbuild/BUILD/git-2.14.2/t/trash%20directory.t9128-git-svn-cmd-branch/svnrepo/branches/a...
    Found possible branch point: file:///home/tmz/rpmbuild/BUILD/git-2.14.2/t/trash%20directory.t9128-git-svn-cmd-branch/svnrepo/trunk => file:///home/tmz/rpmbuild/BUILD/git-2.14.2/t/trash%20directory.t9128-git-svn-cmd-branch/svnrepo/branches/a, 2
    Found branch parent: (refs/remotes/origin/a) bb43a88ce94511096d7d774f4d5feae281603eb0
    Following parent with do_switch
    Successfully followed parent
    r3 = 2e27ee45dadc098239114bc13a6ae6cf6122d16f (refs/remotes/origin/a)
    error: git-svn died of signal 11
    + test_eval_ret_=139
    + want_trace
    + test t = t
    + test t = t
    + set +x
    error: last command exited with $?=139
    not ok 3 - git svn branch tests

The failure in t9167-git-svn-cmd-branch-subproject is similar.