Rename %gitcoredir to %gitexecdir; upstream uses the latter
Using the same variable make it clearer that we're referring to the same
thing as upstream when the Makefile uses gitexecdir.

Remove a stale conditional in git-cvs for handling the case where git's
binaries are stored in %{_bindir}.  This support was dropped after
EL-5's support ended.  Most similar conditionals were removed in 903d8f3
(Remove EL-5 and old Fedora conditionals, 2017-07-22).

Also remove the conditional setting of 'gitexecdir = %{_bindir}' which
was only used on EL-5 to prevent moving the git binaries during the
update to git-1.6.