Add tcl/tk BuildRequires
The git-gui build tries to use tclsh to generate an index file.  It
falls back to a static list, which is why this wasn't noticed before.
It's not a real problem, but git may soon also try to test that wish is
installed before building and installing git-gui and gitk, so this helps
prepare for that future.