Run tests in parallel by default on Fedora
An option to enable parallel test runs was added in 6dc6285 ("Improve
test suite coverage", 2017-11-10).  After further improvements it seems
to be reliable for use in koji builds.

It is still disabled for EPEL builds because make lacks support for the
-O (--output-sync) option.  Without this option the output is much less
readable.

In the future we may be able to move to using 'prove' as the default
test harness which might allow us to run in parallel without requiring
the -O option.