Explicitly enable tests which may be skipped opportunistically
Ensure these tests are not skipped if some dependency or other condition
is not met.  We prefer to see the failure(s) in order to fix the
underlying issue(s).

Disable t9115-git-svn-dcommit-funky-renames on ppc-arches because it
frequently fails.  The port it uses (9115) is already in use.  It is unclear if
this is due to an issue in the test suite or a conflict with some other
process on the build host.

Add mod_dav_svn BuildRequires for t9115-git-svn-dcommit-funky-renames
(GIT_SVN_TEST_HTTPD).