Update BuildRequires for tests
The acl and perl(HTTP::Date) requirements are pulled in by other
packages on Fedora so they were not noticed when the test requirements
were updated in 6dc6285 ("Improve test suite coverage", 2017-11-10).

The highlight package is available in EPEL for EL-6 (on all supported
architectures).  Extend the conditional to only exclude it on ppc64 for
EL-7.