gitk: add Requires: git-gui (#1765113)
While this could arguably be a Recommends: rather than Requires:, we
chose the latter for a few reasons.  The user experience when running
gitk and selecting "Start git gui" from the menu is quite poor.  No
indication is shown to the user graphically.  The only hint as to why
git gui did not start is output to stdout (and is not terribly helpful
for users who may be using gitk and git-gui because they are unfamiliar
with the command-line).

There are no additional dependencies pulled in by git-gui which are not
already dependencies of gitk.  And the git-gui package is relatively
small.

Lastly, the default behavior of Recommends: is the same as Requires: at
this time.

If/when any of these things change, we may revisit whether moving to
Recommends: makes more sense.

Thanks to Vasiliy Glazov and Pavel Cahyna for reporting the issue and
helping to determine the proper resolution.