Use make_build macro when running tests
This will make it possible for buildroots to inject arguments to
make by redefining the %__make macro.

For example, the test target uses gcc to compile fuzz-commit-graph.c, so
one thing this change will allow us to do is to pass CC=clang to make if
we want to try to build with clang.